Move readline into CLI directory
